+++ include/my_user.h
@@ -27,8 +27,8 @@
 C_MODE_START
 
 int parse_user(const char *user_id_str, size_t user_id_len,
-               char *user_name_str, size_t *user_name_len,
-               char *host_name_str, size_t *host_name_len);
+               char *user_name_str, uint *user_name_len,
+               char *host_name_str, uint *host_name_len);
 
 C_MODE_END
