+++ sql-common/my_user.c
@@ -37,14 +37,14 @@
 */
 
 int parse_user(const char *user_id_str, size_t user_id_len,
-               char *user_name_str, size_t *user_name_len,
-               char *host_name_str, size_t *host_name_len)
+               char *user_name_str, uint *user_name_len,
+               char *host_name_str, uint *host_name_len)
 {
   char *p= strrchr(user_id_str, '@');
 
   if (!p)
   {
-    *user_name_len= user_id_len;
+    *user_name_len= (uint)user_id_len;
     *host_name_len= 0;
   }
   else
